"Refactor layout by extracting Sidebar and Topbar components for reusability."
