+++ frontend/src/lib/components/sidebar.svelte
@@ -1,0 +1,85 @@
+<script lang="ts">
+	// Icons importieren
+	import {
+		ChartLine,
+		Files,
+		MessageSquare,
+		Settings,
+		User,
+		GraduationCap,
+		LayoutDashboard
+	} from 'lucide-svelte';
+
+	// Zustand für Sidebar
+	// Funktion zum Umschalten des Sidebar-Zustands
+</script>
+
+<!-- Sidebar -->
+<div
+		class="preset-filled-primary-100-900 flex h-screen w-full flex-col items-center justify-start overflow-hidden rounded transition-all duration-300"
+		class:isFullWidth={$isExpanded}
+		style="width: {$isExpanded ? '200px' : '4rem'}"
+>
+	<!-- Expand/Collapse Button -->
+
+	<div>
+		<!-- Logo -->
+		<p class="mt-3 flex items-center" aria-label="Lernello">
+			<GraduationCap size={24} />
+			{#if $isExpanded}<span class="ml-3">Lernello</span>{/if}
+		</p>
+
+		<!-- Buttons Section -->
+		<div class="mt-3 flex flex-col items-center border-t">
+			<a
+					class="mt-2 flex h-12 w-full items-center rounded"
+					href="/dashboard"
+					aria-label="Dashboard"
+			>
+				<LayoutDashboard size={24} />
+				{#if $isExpanded}<span class="ml-3">Dashboard</span>{/if}
+			</a>
+			<a
+					class="mt-2 flex h-12 w-full items-center rounded"
+					href="/learningkit"
+					aria-label="Learning Kit"
+			>
+				<Files size={24} />
+				{#if $isExpanded}<span class="ml-3">Learning Kits</span>{/if}
+			</a>
+			<a
+					class="mt-2 flex h-12 w-full items-center rounded"
+					href="/statistics"
+					aria-label="Statistics"
+			>
+				<ChartLine size={24} />
+				{#if $isExpanded}<span class="ml-3">Statistics</span>{/if}
+			</a>
+		</div>
+
+		<div class="mt-2 flex flex-col items-center border-t">
+			<a class="mt-2 flex h-12 w-full items-center rounded" href="/settings" aria-label="Settings">
+				<Settings size={24} />
+				{#if $isExpanded}<span class="ml-3">Settings</span>{/if}
+			</a>
+			<a
+					class="relative mt-2 flex h-12 w-full items-center justify-center rounded"
+					href="/notifications"
+					aria-label="Notifications"
+			>
+				<MessageSquare size={24} />
+				<span class="absolute top-0.5 left-2.5 mt-2 ml-2 h-2 w-2 rounded-full bg-indigo-500"></span>
+				{#if $isExpanded}<span class="ml-3">Notifications</span>{/if}
+			</a>
+		</div>
+	</div>
+
+	<a
+			class="mt-auto flex h-16 w-full items-center justify-center"
+			href="/profile"
+			aria-label="Profile"
+	>
+		<User size={24} />
+		{#if $isExpanded}<span class="ml-3">Profile</span>{/if}
+	</a>
+</div>
